arch: make dxvk optional, make softfx default when not present

+++ install_script.sh
@@ -99,6 +99,12 @@ true
 Flycast Dojo (latest prerelease)
 Downloads &amp; installs the latest Flycast Dojo prerelease. Includes script to switch between Flycast Dojo versions used by Fightcade.' > "${TMPDIR}/install_options.list"
 
+if type pacman &> /dev/null; then
+	echo -e 'true
+DXVK
+Vulkan-based implementation of D3D9, D3D10 and D3D11 for Linux / Wine' >> "${TMPDIR}/install_options.list"
+fi
+
 echo -e 'false
 Street Fighter III: 3rd Strike (Grouflon)
 FBNeo Lua Training Scripts
@@ -110,6 +116,7 @@ PLUG=`date +%s`
 
 INSTALL_JOY=false
 INSTALL_DOJO=false
+INSTALL_DXVK=false
 INSTALL_3S_LUA=false
 INSTALL_VF4FT_LUA=false
 
@@ -173,6 +180,13 @@ case "$TAB1" in
 	;;
 esac
 
+case "$TAB1" in
+	*"TRUE|DXVK"*)
+	INSTALL_DXVK=true
+	TOTALSTEPS=$(($TOTALSTEPS+1))
+	;;
+esac
+
 case "$TAB2" in
 	*"TRUE|Street Fighter III"*)
 	INSTALL_3S_LUA=true
@@ -223,10 +237,12 @@ if type pacman &> /dev/null; then
 	yay -S --noconfirm rsync wine wine-mono lib32-mpg123 lib32-libxss lib32-libcurl-gnutls libcurl-gnutls libzip miniupnpc lua53 libao lib32-faudio
 	
 	# install libdxvk
-	wget "https://github.com/eclairevoyant.gpg"
-	sudo pacman-key --add eclairevoyant.gpg
-	yay -S --noconfirm lib32-sdl2 lib32-vulkan-icd-loader libdxvk
-	rm eclairevoyant.gpg
+	if [[ "$INSTALL_DXVK" == "true" ]]; then
+		wget "https://github.com/eclairevoyant.gpg"
+		sudo pacman-key --add eclairevoyant.gpg
+		yay -S --noconfirm lib32-sdl2 lib32-vulkan-icd-loader libdxvk
+		rm eclairevoyant.gpg
+	fi
 elif type apt &> /dev/null; then
 	. /etc/os-release
 
@@ -267,6 +283,11 @@ rm Fightcade-linux-latest.tar.gz
 # set fbneo audio default to xaudio2
 cd $FC_DIR
 sed -i "40ised -i \'s/nAudSelect 0/nAudSelect 1/\' ./emulator/fbneo/config/fcadefbneo.ini" Fightcade2.sh
+
+#set fbneo video to softfx by default when dxvk is not installed
+if [[ "$INSTALL_DXVK" == "false" ]]; then
+	sed -i "41ised -i \'s/nVidSelect 4/nVidSelect 2/\' ./emulator/fbneo/config/fcadefbneo.ini" Fightcade2.sh
+fi
 
 # create symbolic links for included flycast dojo binary
 sudo ln -s "/usr/lib/libzip.so" "/usr/lib/libzip.so.4"
